fix: Carbon addSeconds/addDays int 타입 캐스팅
config()에서 반환되는 문자열을 int로 캐스팅하여
Carbon 에러 수정

+++ app/Providers/AppServiceProvider.php
@@ -21,8 +21,8 @@ class AppServiceProvider extends ServiceProvider
     public function boot(): void
     {
         // Passport token expiration
-        Passport::tokensExpireIn(now()->addSeconds(config('passport.tokens_expire_in', 3600)));
-        Passport::refreshTokensExpireIn(now()->addDays(config('passport.refresh_tokens_expire_in', 14)));
-        Passport::personalAccessTokensExpireIn(now()->addSeconds(config('passport.tokens_expire_in', 3600)));
+        Passport::tokensExpireIn(now()->addSeconds((int) config('passport.tokens_expire_in', 3600)));
+        Passport::refreshTokensExpireIn(now()->addDays((int) config('passport.refresh_tokens_expire_in', 14)));
+        Passport::personalAccessTokensExpireIn(now()->addSeconds((int) config('passport.tokens_expire_in', 3600)));
     }
 }
